% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_split.R
\name{dl_split}
\alias{dl_split}
\title{Worst-Pair-Fitting Damerau Levenshtein Distance}
\usage{
dl_split(a, b, penalty = 0.1)
}
\arguments{
\item{a}{\code{[character]} vector.}

\item{b}{\code{[character]} vector.}

\item{penalty}{\code{[numeric]} scalar. Penalty for using the worst-fit-pair instead of whole string.}
}
\description{
Split strings at spaces and find the distance of the worst fitting pair.
}
\section{Details}{

}

\examples{

dl_split(a,b)

}
